--- test/cache/__init__.py
@@ -1,1 +1,0 @@
-from .cache import save, load
