+++ test/test_methods.py
@@ -1,4 +1,8 @@
 import utils
+import utils.docker
+import utils.name
+import utils.warnings
+
 import parameterized
 import openproblems
