+++ test/test_metrics.py
@@ -1,4 +1,5 @@
-import utils
+import utils.name
+import utils.warnings
 
 import parameterized
 import openproblems
